refactor code,corrected import errors
